[typing] fix isConstant test for global initialization

+++ tests/syntax/ko_global.c
@@ -22,7 +22,7 @@ const int f = f;
 
 int g = g;
 
-int h = (int) &h;
+short h = (short) &h;
 
 const int i;
 int j = i;
